Got basic multiplayer setup working
Implemented a basic multiplayer system in which the user can run two or more instances of the simulator and they will communicate with each other. However, the current limitation is that the robot files loaded must be identical on each machine. In addition, the simulation can look jittery for the clients, so some lag compensation has to be done.