Started on collision handling.
Started on a collision handling system in which the server has priority in calculating the result of the collision. This current system works fairly well, but sometimes when the server initiates collision some weird results can occur. I'm planning on solving this by allowing the server to keep priority until the two robots exceed a certain distance from each other.